I suggest that this is what we want to test
In `'"he was called \"jake\""'` the `\"` essentially escapes the quote
and is replaced by just a quote thus being === to `'"he was called "jake""'`
which cannot be tokenized.

I think what was intended is actually `"he was called \"jake\""` 
and we must then escape the backslash, not the quote.

Because the included string contains single quotes, I don't think we should
change them and it should return the exact same string.

**This breaks the tests**